+++ Engine/Source/Renderer/Light/Light.cpp
@@ -1,0 +1,16 @@
+#include "Light.h"
+
+namespace Core
+{
+    Light::Light()
+    {
+    }
+
+    Light::~Light()
+    {
+    }
+
+    void Light::Render()
+    {
+    }
+}