+++ Engine/Source/Renderer/Light/Light.h
@@ -1,0 +1,16 @@
+#pragma once
+
+#include "Base.h"
+
+namespace Core
+{
+    class Light
+    {
+    public:
+        Light();
+        ~Light();
+
+        virtual void Render();
+    };
+
+}