+++ Engine/Source/Scene/SceneEnvironment.h
@@ -2,6 +2,7 @@
 
 #include "Base.h"
 #include "Renderer/Sky/Sky.h"
+#include "Renderer/Light/DirectionalLight.h"
 
 #include <string>
 
@@ -10,6 +11,7 @@ namespace Core
     struct SceneEnvironment
     {
         Sky *SkyInst;
+        DirectionalLight DirectionalLight;
 
         void Setup();
         void From(SceneEnvironment* env);